Be More Defensive of Incompatible FROST Keys
Because of how `ecrecover` is implemented, not all FROST public keys
are supported by the `ecmulmuladd` trick that we use in our verifier.
While the chance of it happening is exceedingly rare (~1/2**128), just
to be safe, I modified the CLI to check for key validity for some
operations (such as splitting and encoding public key information for
account deployment).